* getopt.c (_getopt_internal): Cast the return value of strlen, which is unsigned, before comparing it with the difference between two pointers, which is unsigned.
CVS-ID: lib/getopt.c 1.3

+++ lib/getopt.c
@@ -498,7 +498,7 @@ _getopt_internal (argc, argv, optstring, longopts, longind, long_only)
       for (p = longopts, option_index = 0; p->name; p++, option_index++)
 	if (!strncmp (p->name, nextchar, nameend - nextchar))
 	  {
-	    if (nameend - nextchar == strlen (p->name))
+	    if (nameend - nextchar == (int) strlen (p->name))
 	      {
 		/* Exact match found.  */
 		pfound = p;
